+++ src/components/inputs.ts
@@ -1,25 +1,30 @@
+import { keys } from 'ts-transformer-keys';
+import { JSX } from '@ionic/core';
 import { defineInput } from '../utils';
 
-export const IonCheckbox = /*@__PURE__*/defineInput(
+export const IonCheckbox = /*@__PURE__*/defineInput<JSX.IonCheckbox>(
   'IonCheckbox',
   'ion-checkbox',
+  keys<JSX.IonCheckbox>(),
   'onIonChange',
   'checked'
 );
-export const IonDatetime = /*@__PURE__*/defineInput('IonDatetime', 'ion-datetime');
-export const IonInput = /*@__PURE__*/defineInput('IonInput', 'ion-input', 'onIonInput');
-export const IonRadio = /*@__PURE__*/defineInput('IonRadio', 'ion-radio', 'onIonSelect');
-export const IonRange = /*@__PURE__*/defineInput('IonRange', 'ion-range');
-export const IonSearchbar = /*@__PURE__*/defineInput(
+export const IonDatetime = /*@__PURE__*/defineInput<JSX.IonDatetime>('IonDatetime', 'ion-datetime', keys<JSX.IonDatetime>());
+export const IonInput = /*@__PURE__*/defineInput<JSX.IonInput>('IonInput', 'ion-input', keys<JSX.IonInput>(), 'onIonInput');
+export const IonRadio = /*@__PURE__*/defineInput<JSX.IonRadio>('IonRadio', 'ion-radio', keys<JSX.IonRadio>(), 'onIonSelect');
+export const IonRange = /*@__PURE__*/defineInput<JSX.IonRange>('IonRange', 'ion-range', keys<JSX.IonRange>());
+export const IonSearchbar = /*@__PURE__*/defineInput<JSX.IonSearchbar>(
   'IonSearchbar',
   'ion-searchbar',
+  keys<JSX.IonSearchbar>(),
   'onIonInput'
 );
-export const IonSelect = /*@__PURE__*/defineInput('IonSelect', 'ion-select');
-export const IonTextarea = /*@__PURE__*/defineInput('IonTextarea', 'ion-textarea');
-export const IonToggle = /*@__PURE__*/defineInput(
+export const IonSelect = /*@__PURE__*/defineInput<JSX.IonSelect>('IonSelect', 'ion-select', keys<JSX.IonSelect>());
+export const IonTextarea = /*@__PURE__*/defineInput<JSX.IonTextarea>('IonTextarea', 'ion-textarea', keys<JSX.IonTextarea>());
+export const IonToggle = /*@__PURE__*/defineInput<JSX.IonToggle>(
   'IonToggle',
   'ion-toggle',
+  keys<JSX.IonToggle>(),
   'onIonChange',
   'checked'
 );