+++ src/components/overlays.ts
@@ -4,21 +4,28 @@ import {
   PopoverOptions,
   actionSheetController,
   modalController,
-  popoverController
+  popoverController,
 } from '@ionic/core';
+import { keys } from 'ts-transformer-keys';
 import { OverlayType, defineOverlay } from '../utils';
 
 export const IonModal = /*@__PURE__*/defineOverlay<HTMLIonModalElement, ModalOptions>(
   OverlayType.Modal,
-  modalController
+  modalController,
+  keys<Omit<ModalOptions, 'component' | 'componentProps' | 'delegate'>>(),
 );
 
 export const IonActionSheet = /*@__PURE__*/defineOverlay<
   HTMLIonActionSheetElement,
   ActionSheetOptions
->(OverlayType.ActionSheet, actionSheetController);
+>(
+  OverlayType.ActionSheet,
+  actionSheetController,
+  keys<ActionSheetOptions>(),
+);
 
 export const IonPopover = /*@__PURE__*/defineOverlay<HTMLIonPopoverElement, PopoverOptions>(
   OverlayType.Popover,
-  popoverController
+  popoverController,
+  keys<Omit<PopoverOptions, 'component' | 'componentProps' | 'delegate'>>(),
 );