+++ src/utils/defineContainer.ts
@@ -1,14 +1,11 @@
 import { FunctionalComponent, h } from 'vue';
 
-type Data = {
-  [key: string]: unknown;
-};
-
-export const defineContainer = (name: string) => {
-  const Container: FunctionalComponent<Data> = (props, { slots }) =>
+export const defineContainer = <Props extends object>(name: string, componentProps: string[]) => {
+  const Container: FunctionalComponent<Props> = (props, { slots }) =>
     h(name, props, slots.default && slots.default());
 
   Container.displayName = name;
+  Container.props = componentProps;
 
   return Container;
 };